requirements.txt
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.11.1